TarefaDomain e TarefaController

+++ src/main/java/com/marcelo/system/domain/PessoaDomain.java
@@ -13,24 +13,24 @@ public class PessoaDomain implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String nome;
 
 	public PessoaDomain() {
 		super();
 	}
 
-	public PessoaDomain(Integer id, String nome) {
+	public PessoaDomain(Long id, String nome) {
 		super();
 		this.id = id;
 		this.nome = nome;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
